Move addresses into osc input controller
