+++ tests/controller/osc/test_input_controller.py
@@ -4,7 +4,7 @@ import pytest
 from pytest_mock import MockerFixture
 
 from vrchat_io.abc.controller import Controller
-from vrchat_io.controller.osc.input_controller import InputController
+from vrchat_io.controller.osc.input_controller import InputController, Axes, Buttons, logger
 
 
 class TestInputController:
@@ -60,7 +60,7 @@ class TestInputController:
 
         command_and_reset_mocker = mocker.patch.object(controller, "command_and_reset")
         command_and_reset_mocker.side_effect = Exception("test error")
-        logger_mocker = mocker.patch("vrchat_io.osc.input_controller.logger.exception")
+        logger_mocker = mocker.patch.object(logger, "exception")
 
         controller._command_and_reset_with_error_catch(*args)
 
@@ -74,3 +74,38 @@ class TestInputController:
         assert isinstance(thread, threading.Thread)
         thread.join()
         mock.assert_called_once_with(*args)
+
+
+def test_Axes():
+    assert Axes.Vertical == "/input/Vertical"
+    assert Axes.Horizontal == "/input/Horizontal"
+    assert Axes.LookHorizontal == "/input/LookHorizontal"
+    assert Axes.UseAxisRight == "/input/UseAxisRight"
+    assert Axes.GrabAxisRight == "/input/GrabAxisRight"
+    assert Axes.MoveHoldFB == "/input/MoveHoldFB"
+    assert Axes.SpinHoldCwCcw == "/input/SpinHoldCwCcw"
+    assert Axes.SpinHoldUD == "/input/SpinHoldUD"
+    assert Axes.SpinHoldLR == "/input/SpinHoldLR"
+
+
+def test_Buttons():
+    assert Buttons.MoveForward == "/input/MoveForward"
+    assert Buttons.MoveBackward == "/input/MoveBackward"
+    assert Buttons.MoveLeft == "/input/MoveLeft"
+    assert Buttons.MoveRight == "/input/MoveRight"
+    assert Buttons.LookLeft == "/input/LookLeft"
+    assert Buttons.LookRight == "/input/LookRight"
+    assert Buttons.Jump == "/input/Jump"
+    assert Buttons.Run == "/input/Run"
+    assert Buttons.ComfortLeft == "/input/ComfortLeft"
+    assert Buttons.ComfortRight == "/input/ComfortRight"
+    assert Buttons.DropRight == "/input/DropRight"
+    assert Buttons.UseRight == "/input/UseRight"
+    assert Buttons.GrabRight == "/input/GrabRight"
+    assert Buttons.DropLeft == "/input/DropLeft"
+    assert Buttons.UseLeft == "/input/UseLeft"
+    assert Buttons.GrabLeft == "/input/GrabLeft"
+    assert Buttons.PanicButton == "/input/PanicButton"
+    assert Buttons.QuickMenuToggleLeft == "/input/QuickMenuToggleLeft"
+    assert Buttons.QuickMenuToggleRight == "/input/QuickMenuToggleRight"
+    assert Buttons.Voice == "/input/Voice"
