--- vrchat_io/controller/osc/addresses.py
@@ -1,43 +1,0 @@
-"""This file contains OSC addresses for the VRChat.
-
-See official docs: https://docs.vrchat.com/docs/osc-overview OSC as Input
-Controller:
-"""
-
-# OSC addresses for input controller.
-# See official docs: https://docs.vrchat.com/docs/osc-as-input-controller
-
-
-class Axes:
-    Vertical = "/input/Vertical"
-    Horizontal = "/input/Horizontal"
-    LookHorizontal = "/input/LookHorizontal"
-    UseAxisRight = "/input/UseAxisRight"
-    GrabAxisRight = "/input/GrabAxisRight"
-    MoveHoldFB = "/input/MoveHoldFB"
-    SpinHoldCwCcw = "/input/SpinHoldCwCcw"
-    SpinHoldUD = "/input/SpinHoldUD"
-    SpinHoldLR = "/input/SpinHoldLR"
-
-
-class Buttons:
-    MoveForward = "/input/MoveForward"
-    MoveBackward = "/input/MoveBackward"
-    MoveLeft = "/input/MoveLeft"
-    MoveRight = "/input/MoveRight"
-    LookLeft = "/input/LookLeft"
-    LookRight = "/input/LookRight"
-    Jump = "/input/Jump"
-    Run = "/input/Run"
-    ComfortLeft = "/input/ComfortLeft"
-    ComfortRight = "/input/ComfortRight"
-    DropRight = "/input/DropRight"
-    UseRight = "/input/UseRight"
-    GrabRight = "/input/GrabRight"
-    DropLeft = "/input/DropLeft"
-    UseLeft = "/input/UseLeft"
-    GrabLeft = "/input/GrabLeft"
-    PanicButton = "/input/PanicButton"
-    QuickMenuToggleLeft = "/input/QuickMenuToggleLeft"
-    QuickMenuToggleRight = "/input/QuickMenuToggleRight"
-    Voice = "/input/Voice"
